Speaker and Attendee UI along with their corresponding controller partially finished

+++ phase1/src/Entity/Room.java
@@ -22,7 +22,7 @@ public class Room {
     }
 
     // Get the maximum capacity of the room.
-    public double getMaxCapacity() {
+    public Integer getMaxCapacity() {
         return MaxCapacity;
     }
